requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
scikit-learn==1.8.0
pandas==2.2.2
numpy==2.1.1
joblib==1.4.2
pydantic-settings==2.6.1
